Separates IconTransform of the Icon component
