+++ lib/components/Icon.js
@@ -2,44 +2,30 @@ import {InvalidArgument, InvalidElement, RuntimeException} from "../errors.js";
 import {AriaAttr, CustomAttr} from "../misc/attributes.js";
 import ns from "../misc/namespaces.js";
 import snippets from "./snippets.js";
+import IconTransform from "./IconTransform.js";
 import Pictogram from "./Pictogram.js";
-import Transform from "./Transform.js";
 
-class IconTransform extends Transform {
 
-  // Moves the SVG element from its initial position.
-  // Ovverides the parent's method for use stylesheets.
-  translate(x , y) {
-    const style = this.svgElement.style;
-    let tfmStr = style.transform || style.msTransform ||
-                 style.webkitTransform || style.oTransform || "";
-    if (tfmStr) {
-      tfmStr += " ";
-    }
-
-    tfmStr += "translate(" + [x, y].join(",") + ")";
-
-    style.msTransform = tfmStr;
-    style.oTransform = tfmStr;
-    style.webkitTransform = tfmStr;
-    style.transform = tfmStr;
-  }
-
-}
-
-// @private {IconTransform} tfmInstance - Instance of the Transform class
-// const tfmInstance = Symbol("tfmInstance");
-
+// Private attributes
 const _aria = new WeakMap();
 const _connector = new WeakMap();
 const _element = new WeakMap();
 const _meta = new WeakMap();
+const _pictogram = new WeakMap();
 const _svgElement = new WeakMap();
 const _tfmInstance = new WeakMap();
 
-
+/**
+ * Represents a SVG icon item, which is composed by a declarative HTML Element
+ * and by a generated SVGElement linked to a {@link Pictogram}.
+ */
 class Icon {
 
+  /**
+   * Constructor of the object.
+   * @param {!HTMLElement} element - The element used to report the location of the icon from the document.
+   * @throws {InvalidElement} When the element argument is wrong.
+   */
   constructor(element) {
     if (!(element instanceof HTMLElement)) {
       throw new InvalidElement("HTMLElement", element);
@@ -55,6 +41,7 @@ class Icon {
     _connector.set(this, connector);
     _element.set(this, element);
     _meta.set(this, new CustomAttr(element));
+    _pictogram.set(this, Pictogram.replacement);
     _svgElement.set(this, svgElement);
     // _tfmInstance.set(this, null);
 
@@ -62,14 +49,17 @@ class Icon {
       aria: { get: function() {
         return _aria.get(this);
       }},
+      connector: { get: function() {
+        return _connector.get(this);
+      }},
       element: { get: function() {
         return _element.get(this);
       }},
       meta: { get: function() {
         return _meta.get(this);
       }},
-      connector: { get: function() {
-        return _connector.get(this);
+      pictogram: { get: function() {
+        return _pictogram.get(this);
       }},
       svgElement: { get: function() {
         return _svgElement.get(this);
@@ -77,6 +67,10 @@ class Icon {
     });
   }
 
+  /**
+   * Applies the changes provided by the customized attributes of the element
+   * @example <span data-icon="foo" data-rotate="45"></span>
+   */
   applyTransforms() {
     if (this.meta.has("rotate")) {
       this.transform().rotate(this.meta.get("rotate"));
@@ -85,12 +79,16 @@ class Icon {
     }
   }
 
+  /**
+   * Displays the icon in the document.
+   * @param {!Pictogram} pictogram - The pictogram to be displayed by the icon.
+   */
   draw(pictogram) {
     if (!(pictogram instanceof Pictogram)) {
       throw new InvalidArgument("Invalid Pictogram");
     }
     // this.erase();
-    this.pictogram = pictogram;
+    _pictogram.set(this, pictogram);
 
     // Adds DOM nodes required for rendering
     this.element.appendChild(this.svgElement);
@@ -120,32 +118,51 @@ class Icon {
     this.applyTransforms();
   }
 
+  /**
+   * Cleans the component before its destruction.
+   */
   remove() {
     // Empties the element of its children.
     const elem = this.element;
     while(elem.firstChild) {
       elem.removeChild(elem.firstChild);
     }
+
+    // Deletes the attributes generated by the component.
     this.aria.remove("labelledby");
     this.aria.remove("descriptedby");
 
+    // Purges the private attributes.
     _aria.delete(this);
     _connector.delete(this);
     _element.delete(this);
     _meta.delete(this);
+    _pictogram.delete(this);
     _svgElement.delete(this);
     _tfmInstance.delete(this);
   }
 
+  /**
+   * Returns the size infos of the icon.
+   * @returns {Object} An object with the size and its position of the icon.
+   */
   getSize() {
     return this.svgElement.getBoundingClientRect();
   }
 
+  /**
+   * Returns the icon reference size
+   * @returns {number} Corresponds to the longer side of the icon.
+   */
   getRefSize() {
     const size = this.getSize();
     return Math.max(size.height, size.width);
   }
 
+  /**
+   * Loads a Transform instance of the icon.
+   * @returns {IconTransform} The transform object of the icon.
+   */
   transform() {
     let tfm = _tfmInstance.get(this);
     if (!tfm) {
@@ -160,6 +177,10 @@ class Icon {
 
 }
 
+/**
+ * Snippets used by the component for draw the icon.
+ * @static
+ */
 Icon.snippets = snippets;
